% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_results.R
\name{clean_results}
\alias{clean_results}
\title{Clean the resulting dataframe to prep it for modelling. Add max depth column.}
\usage{
clean_results(u, d)
}
\arguments{
\item{u}{User results from ShinyApp. Use r$bottom_phase_data}

\item{d}{Full data frame. Use r$full_data}
}
\value{
u but modified with extra columns.
}
\description{
Clean the resulting dataframe to prep it for modelling. Add max depth column.
}
