% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResults.R
\name{loadResults}
\alias{loadResults}
\title{Loading the data produced from the Shiny App}
\usage{
loadResults(outputDir, full_data)
}
\arguments{
\item{outputDir}{A string containing the path to where the ShinyApp data has been stored.}
}
\value{
A dataframe containing the start and end times of the dive as well as the start
and end times of the user-defined bottom phase. The dive number and the time stamp of when
the bottom phase data had been saved are included as well.
}
\description{
Loading the data produced from the Shiny App
}
\examples{
loadResults("path/to/Shiny/data/ouputs/")
}
