% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_max_depth.R
\name{percent_max_depth}
\alias{percent_max_depth}
\title{Produce Optimized Theta value}
\usage{
percent_max_depth(full_data, bottom_phase_data, start = TRUE)
}
\arguments{
\item{d}{dataframe of results from Shiny App.}
}
\value{
Optimized Theta
}
\description{
Produce Optimized Theta value
}
