% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Computes summary statistics for the previously created phases dataset.}
\usage{
results(f, outdir)
}
\arguments{
\item{phases}{A dataframe corresponding to the output of process_data$phases.}
}
\value{
A dataframe of the dives_by_row. Calling summary on this object will
print basic summary statistics about the data.
}
\description{
Computes summary statistics for the previously created phases dataset.
}
\examples{
results(full_data, "~/Documents/whale_data/")
}
