% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBottomPhaseApp.R
\name{runBottomPhaseApp}
\alias{runBottomPhaseApp}
\title{Run the ShinyApp with the user's own dataset.}
\usage{
runBottomPhaseApp(
  entire_record,
  file_path,
  init_num = 10,
  custom_dives = c(),
  rand_select = TRUE,
  continuing = FALSE,
  launch_browser = FALSE
)
}
\arguments{
\item{entire_record}{A dataframe containing time-depth data as produced from the read_dive_data function.}

\item{file_path}{A string indicating a path to a folder where the resulting data will be stored and accessed.}

\item{init_num}{The initial number of randomly selected dives that the user wishes to label the bottom phase.
Defaults to 10 dives.}

\item{rand_select}{Defaults to TRUE. If FALSE, dives will not be randomly chosen from the dataset, but instead will
start at 1 and continue until init_num is reached.}

\item{continuing}{A boolean, defaulting to FALSE. If you are continuing bottom phase identification from a
previous session, and you have your previous data stored in file_path, then set this parameter to be TRUE.
The app will ignore all dives which have been previously labelled, allowing you to "pick up where you left off".
If this parameter is set to be TRUE, then you need to make sure that the inputted dataframes are identical to
the ones from the previous session. The data should come from the same tagged animal, and all other parameters should be
the same as they were before, including mindepth and surface from the find_dives function. If there are discrepancies or
differences between the previous sessions' data and the current sessions' data, errors can potentially arise since the
time stamps of the dives in the current dataframe may not line up with the ones that are currently saved on your machine.}

\item{individual_dives}{A dataframe wherein each row is an individual dive from the animal. This
must be equivalent to the output of bustR::find_dives().}
}
\value{
Runs an interactive ShinyApp and saves your responses as .csv files in the file_path folder.
}
\description{
Run the ShinyApp with the user's own dataset.
}
\examples{
runBottomPhaseApp(dives$full_data, init_num = 50, file_path = "~/Desktop/saved_data_whale1/", continuing = TRUE)
runRBottomPhaseApp(dives$full_data, init_num = 10, file_path = "C://Users/name/Documents/whale_data/", continuing = FALSE)
}
